requirements.txt
langchain>=0.0.267
langchain-community>=0.0.1
chromadb>=0.4.13
streamlit>=1.26.0
apache-flink>=1.17.0
pandas>=2.0.3
numpy>=1.24.3
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
dashscope>=1.13.6
pydantic>=2.4.2
tiktoken>=0.5.1
beautifulsoup4
requests
selenium
webdriver-manager